Move the reducers to the app package
