+++ src/test/kotlin/com/larastudios/chambrier/app/DiscoveredDevicesReducerTest.kt
@@ -1,5 +1,8 @@
-package com.larastudios.chambrier.app.domain
+package com.larastudios.chambrier.app
 
+import com.larastudios.chambrier.app.domain.DiscoveredDevices
+import com.larastudios.chambrier.app.domain.Event
+import com.larastudios.chambrier.app.domain.State
 import com.larastudios.chambrier.lightDevice
 import com.larastudios.chambrier.lightDevice2
 import com.larastudios.chambrier.renamedLightDevice
